+++ docs/CHANGELOG.md
@@ -14,10 +14,12 @@
 - added admin-db-commands
 - fixed import for mysql
 - use commons 5.5.0
+- added media-import-commands
  
 ### new features
 - frontend/backend: added sectionDetails
 - backend: added admin-db-command
+- backend: added media-import-commands
 
 ### improvements
 - frontend: added additional groupby track+trip+location for image/track/trip/video/route-list